test: split tooling_skills integration monolith into domain modules (#1512)
